Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/eventsource/format-field-id-2-expected.txt


FAIL EventSource: Last-Event-ID (2) assert_equals: expected "hello" but got "…"
